// components/runtime/src/ts/generated/internal/IEventInviteCreate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { IInviteTargetUser } from "../discord/IInviteTargetUser";
import type { InviteTargetType } from "../discord/InviteTargetType";
import type { IUser } from "./IUser";

export type IEventInviteCreate = {
  channelId: string;
  code: string;
  createdAt: number;
  inviter?: IUser;
  maxAge: number;
  maxUses: number;
  targetUserType?: InviteTargetType;
  targetUser?: IInviteTargetUser;
  temporary: boolean;
  uses: number;
};