// components/runtime/src/ts/generated/internal/IEventVoiceStateUpdate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { IMember } from "./Member";

export interface IEventVoiceStateUpdate {
  channelId: string | null;
  deaf: boolean;
  guildId: string;
  member?: IMember;
  mute: boolean;
  selfDeaf: boolean;
  selfMute: boolean;
  selfStream: boolean;
  selfVideo: boolean;
  sessionId: string;
  suppress: boolean;
  userId: string;
  requestToSpeakTimestamp: number | null;
}